vibe-coding-lab/02_Cursor_CurrencyTracker/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
rich>=13.0.0
